Correcting select option event to avoid close bug
When the view area is shortest than the dropdown height, clicking on the select2 causes the scroll position of the page to change (because the search input is focused), this position change causes the cursor (still on mousedown) to go over an option and than the mouseup event selects this option and closes the select2. So you open and close it in a fraction of second selecting an option that you don't want. This keeps happening all the time.